requirements.txt
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
nexmo==1.5.0
pycparser==2.19
PyJWT==1.7.1
python-dotenv==0.6.4
pytz==2019.3
requests==2.22.0
six==1.13.0
urllib3==1.25.7
visitor==0.1.3
Werkzeug==0.16.0
